+++ src/frontend/projetos.html
@@ -14,7 +14,7 @@
     
     <title>Projetos</title>
 </head>
-    <body onload="geraTabelaProj();">
+    <body onload="geraTabelaProj();getEmployees();">
         <script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/2.5.0/Chart.min.js"></script>
         <!--PÁGINA DE PROJETOS-->
